Use timer to call the output change callback of the sub-thread.
The output change callback of tdm_server and tdm_vblank was called
in the main thread. And it made the multi thread issue. If we use
the timer, we can call the sub-thread's output change callback in
sub-thread.

And, after tdm_event_loop_deinit, we don't worry about thread things
because it's finalized.

Change-Id: I69013b302673551b887384325b14c94e3a53646c