Use timer to call the output change callback of the sub-thread.
The output change callback of tdm_server and tdm_vblank was called
in the main thread. And it made the multi thread issue. If we use
the timer, we can call the sub-thread's output change callback in
sub-thread.

And, after tdm_event_loop_deinit, we don't worry about thread things
because it's finalized.

Change-Id: I880e76dfe8020621d94c46fc55e98b4e93a3694f